+++ src/admin_ui/src/admin/resources.tsx
@@ -20,6 +20,7 @@ export function getResources(apiInfo: IApiInfo) {
     return (
         apiInfo.mapResources((resourceInfo) => (
             <Resource
+                key={resourceInfo.name}
                 name={resourceInfo.name}
                 options={{label: resourceInfo.title}}
                 {...getResourceViews(resourceInfo)}