+++ src/admin_ui/src/admin/widgets/index.ts
@@ -3,3 +3,4 @@ export * from "./json_input";
 export * from "./mapping_field";
 export * from "./mapping_input";
 export * from "./simple_array_field";
+export * from "./nested_array_field";