+++ src/restfw_admin/restfw_admin/resource_admin.py
@@ -14,7 +14,7 @@ from . import models
 from .fields import get_field_widgets, get_input_widgets
 from .models import FieldModel
 from .typing import ColanderNode
-from .widgets import Widget
+from .widgets import ArrayField, ChipField, MappingField, NestedArrayField, TextField, Widget
 
 
 @dataclasses.dataclass()
@@ -120,6 +120,7 @@ class ResourceAdmin:
                     self.list_view,
                     models.ListViewModel,
                     fields_type='view',
+                    use_nested_array_field=True,
                 )
 
     def get_show_view(self) -> Optional[models.ShowViewModel]:
@@ -166,6 +167,7 @@ class ResourceAdmin:
             view_model_class: Type[models.ViewModelType],
             *,
             fields_type: Literal['view', 'input'],
+            use_nested_array_field=False,
     ) -> Optional[models.ViewModelType]:
         if schema_node is None:
             return
@@ -175,7 +177,7 @@ class ResourceAdmin:
         else:
             get_widgets = get_field_widgets
         widgets = get_widgets(self._registry, schema_node)
-        fields = self._widgets_to_fields(view_settings, widgets)
+        fields = self._widgets_to_fields(view_settings, widgets, use_nested_array_field)
         return view_model_class(fields=fields)
 
     def _get_schema_node(
@@ -195,7 +197,12 @@ class ResourceAdmin:
         if schema_class:
             return schema_class()
 
-    def _widgets_to_fields(self, view_settings: ViewSettings, widgets: Dict[str, Widget]) -> List[FieldModel]:
+    def _widgets_to_fields(
+            self,
+            view_settings: ViewSettings,
+            widgets: Dict[str, Widget],
+            use_nested_array_field=False
+    ) -> List[FieldModel]:
         only_field_names: list[str] = []
         for fields in (self.default_fields, self.fields, view_settings.fields):
             if fields:
@@ -210,17 +217,43 @@ class ResourceAdmin:
         if only_field_names:
             filtered_widgets = {}
             for name in only_field_names:
+                in_array_field = False
                 widget = None
                 cur_widgets = widgets
                 for sub_name in name.split('.'):
                     if cur_widgets is None:
                         break
+                    if isinstance(widget, ArrayField):
+                        in_array_field = True
                     widget = cur_widgets.get(sub_name, None)
                     if widget is None:
                         break
                     cur_widgets = getattr(widget, 'fields', None)
                 else:
                     if widget is not None:
+                        if in_array_field and use_nested_array_field:
+                            if isinstance(widget, MappingField):
+                                widget = NestedArrayField(
+                                    label=widget.label,
+                                    fields=widget.fields,
+                                )
+                            elif isinstance(widget, ArrayField):
+                                widget = NestedArrayField(
+                                    label=widget.label,
+                                    fields=widget.fields,
+                                )
+                            elif isinstance(widget, NestedArrayField):
+                                pass
+                            else:
+                                if isinstance(widget, TextField):
+                                    widget = ChipField(
+                                        label=widget.label,
+                                    )
+                                widget = NestedArrayField(
+                                    label=widget.label,
+                                    fields={'': widget},
+                                    single_field=True,
+                                )
                         filtered_widgets[name] = widget
             widgets = filtered_widgets
         return [